Moving plotting panel out into own class
